% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdb_operations.R
\name{docdb_create_db_instance}
\alias{docdb_create_db_instance}
\title{Creates a new DB instance}
\usage{
docdb_create_db_instance(DBInstanceIdentifier, DBInstanceClass, Engine,
  AvailabilityZone, PreferredMaintenanceWindow, AutoMinorVersionUpgrade,
  Tags, DBClusterIdentifier, PromotionTier)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The DB instance identifier. This parameter is stored as a lowercase
string.

Constraints:
\itemize{
\item Must contain from 1 to 63 letters, numbers, or hyphens.
\item The first character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}

Example: \code{mydbinstance}}

\item{DBInstanceClass}{[required] The compute and memory capacity of the DB instance; for example,
\code{db.m4.large}.}

\item{Engine}{[required] The name of the database engine to be used for this instance.

Valid value: \code{docdb}}

\item{AvailabilityZone}{The Amazon EC2 Availability Zone that the DB instance is created in.

Default: A random, system-chosen Availability Zone in the endpoint's
AWS Region.

Example: \code{us-east-1d}

Constraint: The \code{AvailabilityZone} parameter can't be specified if the
\code{MultiAZ} parameter is set to \code{true}. The specified Availability Zone
must be in the same AWS Region as the current endpoint.}

\item{PreferredMaintenanceWindow}{The time range each week during which system maintenance can occur, in
Universal Coordinated Time (UTC).

Format: \code{ddd:hh24:mi-ddd:hh24:mi}

The default is a 30-minute window selected at random from an 8-hour
block of time for each AWS Region, occurring on a random day of the
week.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.}

\item{AutoMinorVersionUpgrade}{Indicates that minor engine upgrades are applied automatically to the DB
instance during the maintenance window.

Default: \code{true}}

\item{Tags}{The tags to be assigned to the DB instance.}

\item{DBClusterIdentifier}{[required] The identifier of the DB cluster that the instance will belong to.}

\item{PromotionTier}{A value that specifies the order in which an Amazon DocumentDB replica
is promoted to the primary instance after a failure of the existing
primary instance.

Default: 1

Valid values: 0-15}
}
\description{
Creates a new DB instance.
}
\section{Request syntax}{
\preformatted{svc$create_db_instance(
  DBInstanceIdentifier = "string",
  DBInstanceClass = "string",
  Engine = "string",
  AvailabilityZone = "string",
  PreferredMaintenanceWindow = "string",
  AutoMinorVersionUpgrade = TRUE|FALSE,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  DBClusterIdentifier = "string",
  PromotionTier = 123
)
}
}

\keyword{internal}
