% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticache_operations.R
\name{elasticache_delete_cache_parameter_group}
\alias{elasticache_delete_cache_parameter_group}
\title{Deletes the specified cache parameter group}
\usage{
elasticache_delete_cache_parameter_group(CacheParameterGroupName)
}
\arguments{
\item{CacheParameterGroupName}{[required] The name of the cache parameter group to delete.

The specified cache security group must not be associated with any
clusters.}
}
\description{
Deletes the specified cache parameter group. You cannot delete a cache
parameter group if it is associated with any cache clusters.
}
\section{Request syntax}{
\preformatted{svc$delete_cache_parameter_group(
  CacheParameterGroupName = "string"
)
}
}

\keyword{internal}
