% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsdataservice_service.R
\name{rdsdataservice}
\alias{rdsdataservice}
\title{AWS RDS DataService}
\usage{
rdsdataservice()
}
\description{
AWS RDS DataService provides Http Endpoint to query RDS databases.
}
\section{Operations}{

\tabular{ll}{
\link[=rdsdataservice_execute_sql]{execute_sql} \tab Executes any SQL statement on the target database synchronously
}
}

\examples{
\donttest{svc <- rdsdataservice()
svc$execute_sql(
  Foo = 123
)}

}
