% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsdataservice_operations.R
\name{rdsdataservice_execute_sql}
\alias{rdsdataservice_execute_sql}
\title{Executes any SQL statement on the target database synchronously}
\usage{
rdsdataservice_execute_sql(awsSecretStoreArn, database,
  dbClusterOrInstanceArn, schema, sqlStatements)
}
\arguments{
\item{awsSecretStoreArn}{[required] ARN of the db credentials in AWS Secret Store or the friendly secret name}

\item{database}{Target DB name}

\item{dbClusterOrInstanceArn}{[required] ARN of the target db cluster or instance}

\item{schema}{Target Schema name}

\item{sqlStatements}{[required] SQL statement(s) to be executed. Statements can be chained by using semicolons}
}
\description{
Executes any SQL statement on the target database synchronously
}
\section{Request syntax}{
\preformatted{svc$execute_sql(
  awsSecretStoreArn = "string",
  database = "string",
  dbClusterOrInstanceArn = "string",
  schema = "string",
  sqlStatements = "string"
)
}
}

\keyword{internal}
