% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_rotate_encryption_key}
\alias{redshift_rotate_encryption_key}
\title{Rotates the encryption keys for a cluster}
\usage{
redshift_rotate_encryption_key(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier of the cluster that you want to rotate the
encryption keys for.

Constraints: Must be the name of valid cluster that has encryption
enabled.}
}
\description{
Rotates the encryption keys for a cluster.
}
\section{Request syntax}{
\preformatted{svc$rotate_encryption_key(
  ClusterIdentifier = "string"
)
}
}

\keyword{internal}
